requirements.txt
amulet-core
pyinstaller
